third_party/blink/web_tests/html/details_summary/details-click-controls-expected.txt

PASS details.open is false
PASS details.open is false
PASS details.open is false
PASS details.open is false
PASS details.open is true
PASS successfullyParsed is true

TEST COMPLETE
